Tooltip for SetHook codes in debug stream.
